Move tester to test case (#4)
* Rename assistants to agents

* Move tester to test case
