+++ pyproject.toml
@@ -15,7 +15,7 @@ openai = "==1.35.14"
 [tool.poetry.group.dev.dependencies]
 astroid = "==3.2.3"
 jupyterlab = "==4.2.4"
-pytest = "^8.2.2"
+pytest = "==8.2.2"
 
 [build-system]
 requires = ["poetry-core"]
